requirements.txt
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
dash==2.17.1
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_canvas==0.1.0
dnspython==2.6.1
email_validator==2.2.0
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.4
Flask==3.0.3
Flask-Cors==4.0.1
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.0
mdurl==0.1.2
nest-asyncio==1.6.0
networkx==3.3
numpy==2.0.0
openpyxl==3.1.4
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pillow==10.3.0
plotly==5.22.0
protobuf==5.27.1
pyarrow==16.1.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.1
scikit-image==0.24.0
scikit-learn==1.5.0
scipy==1.13.1
setuptools==70.1.0
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
starlette==0.37.2
streamlit==1.36.0
tenacity==8.4.1
threadpoolctl==3.5.0
tifffile==2024.6.18
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
urllib3==2.2.2
uvicorn==0.30.1
watchdog==4.0.1
watchfiles==0.22.0
websockets==12.0
Werkzeug==3.0.3
wheel==0.43.0
zipp==3.19.2
